pyproject.toml
[project]
name = "psy-opd"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "firebase-admin>=6.8.0",
    "httpx>=0.28.1",
    "passlib[bcrypt]>=1.7.4",
    "pydantic>=2.11.5",
    "pytest>=8.3.5",
    "pytest-asyncio>=1.0.0",
    "python-dotenv>=1.1.0",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "pydantic-settings>=2.9.1",
    "google-genai>=1.19.0",
]
